Move internal classes to the internal package
