+++ sentry-android-core/src/main/java/io/sentry/android/core/DefaultAndroidEventProcessor.java
@@ -24,10 +24,10 @@ import io.sentry.ILogger;
 import io.sentry.SentryBaseEvent;
 import io.sentry.SentryEvent;
 import io.sentry.SentryLevel;
-import io.sentry.android.core.util.ConnectivityChecker;
-import io.sentry.android.core.util.DeviceOrientations;
-import io.sentry.android.core.util.MainThreadChecker;
-import io.sentry.android.core.util.RootChecker;
+import io.sentry.android.core.internal.util.ConnectivityChecker;
+import io.sentry.android.core.internal.util.DeviceOrientations;
+import io.sentry.android.core.internal.util.MainThreadChecker;
+import io.sentry.android.core.internal.util.RootChecker;
 import io.sentry.protocol.App;
 import io.sentry.protocol.Device;
 import io.sentry.protocol.OperatingSystem;
