+++ sentry-android-core/src/main/java/io/sentry/android/core/UserInteractionIntegration.java
@@ -9,9 +9,9 @@ import io.sentry.IHub;
 import io.sentry.Integration;
 import io.sentry.SentryLevel;
 import io.sentry.SentryOptions;
-import io.sentry.android.core.gestures.NoOpWindowCallback;
-import io.sentry.android.core.gestures.SentryGestureListener;
-import io.sentry.android.core.gestures.SentryWindowCallback;
+import io.sentry.android.core.internal.gestures.NoOpWindowCallback;
+import io.sentry.android.core.internal.gestures.SentryGestureListener;
+import io.sentry.android.core.internal.gestures.SentryWindowCallback;
 import io.sentry.util.Objects;
 import java.io.Closeable;
 import java.io.IOException;
