+++ sentry-android-core/src/main/java/io/sentry/android/core/internal/gestures/NoOpWindowCallback.java
@@ -1,4 +1,4 @@
-package io.sentry.android.core.gestures;
+package io.sentry.android.core.internal.gestures;
 
 import android.view.ActionMode;
 import android.view.KeyEvent;
