+++ sentry-android-core/src/main/java/io/sentry/android/core/internal/gestures/SentryGestureListener.java
@@ -1,4 +1,4 @@
-package io.sentry.android.core.gestures;
+package io.sentry.android.core.internal.gestures;
 
 import android.view.GestureDetector;
 import android.view.MotionEvent;
