+++ sentry-android-core/src/main/java/io/sentry/android/core/internal/util/DeviceOrientations.java
@@ -1,4 +1,4 @@
-package io.sentry.android.core.util;
+package io.sentry.android.core.internal.util;
 
 import android.content.res.Configuration;
 import io.sentry.protocol.Device;
