+++ sentry-android-core/src/test/java/io/sentry/android/core/gestures/SentryWindowCallbackTest.kt
@@ -11,7 +11,9 @@ import com.nhaarman.mockitokotlin2.never
 import com.nhaarman.mockitokotlin2.verify
 import com.nhaarman.mockitokotlin2.whenever
 import io.sentry.android.core.SentryAndroidOptions
-import io.sentry.android.core.gestures.SentryWindowCallback.MotionEventObtainer
+import io.sentry.android.core.internal.gestures.SentryGestureListener
+import io.sentry.android.core.internal.gestures.SentryWindowCallback
+import io.sentry.android.core.internal.gestures.SentryWindowCallback.MotionEventObtainer
 import org.junit.Test
 
 class SentryWindowCallbackTest {
