+++ sentry-android-core/src/test/java/io/sentry/android/core/gestures/ViewUtilsTest.kt
@@ -6,6 +6,7 @@ import android.view.View
 import com.nhaarman.mockitokotlin2.doReturn
 import com.nhaarman.mockitokotlin2.mock
 import com.nhaarman.mockitokotlin2.whenever
+import io.sentry.android.core.internal.gestures.ViewUtils
 import org.junit.Test
 import kotlin.test.assertEquals
