+++ sentry-android-core/src/test/java/io/sentry/android/core/util/MainThreadCheckerTest.kt
@@ -1,6 +1,7 @@
 package io.sentry.android.core.util
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
+import io.sentry.android.core.internal.util.MainThreadChecker
 import io.sentry.protocol.SentryThread
 import org.junit.runner.RunWith
 import kotlin.test.Test
